requirements.txt
crewai
crewai[tools]
langchain
langchain_openai
python-dotenv
pandas